Split tests from helper code
