+++ tests/vm_test_util.go
@@ -1,10 +1,8 @@
-package vm
+package tests
 
 import (
 	"bytes"
 	"math/big"
-	"os"
-	"path/filepath"
 	"strconv"
 	"testing"
 
@@ -84,12 +82,12 @@ func RunVmTest(p string, t *testing.T) {
 
 	for name, test := range tests {
 		/*
-			vm.Debug = true
-			glog.SetV(4)
-			glog.SetToStderr(true)
-			if name != "Call50000_sha256" {
-				continue
-			}
+		   vm.Debug = true
+		   glog.SetV(4)
+		   glog.SetToStderr(true)
+		   if name != "Call50000_sha256" {
+		     continue
+		   }
 		*/
 		db, _ := ethdb.NewMemDatabase()
 		statedb := state.New(common.Hash{}, db)
@@ -218,171 +216,4 @@ func RunVmTest(p string, t *testing.T) {
 		//fmt.Println(string(statedb.Dump()))
 	}
 	logger.Flush()
-}
-
-// I've created a new function for each tests so it's easier to identify where the problem lies if any of them fail.
-func TestVMArithmetic(t *testing.T) {
-	const fn = "../files/VMTests/vmArithmeticTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestBitwiseLogicOperation(t *testing.T) {
-	const fn = "../files/VMTests/vmBitwiseLogicOperationTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestBlockInfo(t *testing.T) {
-	const fn = "../files/VMTests/vmBlockInfoTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestEnvironmentalInfo(t *testing.T) {
-	const fn = "../files/VMTests/vmEnvironmentalInfoTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestFlowOperation(t *testing.T) {
-	const fn = "../files/VMTests/vmIOandFlowOperationsTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestLogTest(t *testing.T) {
-	const fn = "../files/VMTests/vmLogTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestPerformance(t *testing.T) {
-	const fn = "../files/VMTests/vmPerformanceTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestPushDupSwap(t *testing.T) {
-	const fn = "../files/VMTests/vmPushDupSwapTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestVMSha3(t *testing.T) {
-	const fn = "../files/VMTests/vmSha3Test.json"
-	RunVmTest(fn, t)
-}
-
-func TestVm(t *testing.T) {
-	const fn = "../files/VMTests/vmtests.json"
-	RunVmTest(fn, t)
-}
-
-func TestVmLog(t *testing.T) {
-	const fn = "../files/VMTests/vmLogTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestInputLimits(t *testing.T) {
-	const fn = "../files/VMTests/vmInputLimits.json"
-	RunVmTest(fn, t)
-}
-
-func TestInputLimitsLight(t *testing.T) {
-	const fn = "../files/VMTests/vmInputLimitsLight.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateSystemOperations(t *testing.T) {
-	const fn = "../files/StateTests/stSystemOperationsTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateExample(t *testing.T) {
-	const fn = "../files/StateTests/stExample.json"
-	RunVmTest(fn, t)
-}
-
-func TestStatePreCompiledContracts(t *testing.T) {
-	const fn = "../files/StateTests/stPreCompiledContracts.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateRecursiveCreate(t *testing.T) {
-	const fn = "../files/StateTests/stRecursiveCreate.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateSpecial(t *testing.T) {
-	const fn = "../files/StateTests/stSpecialTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateRefund(t *testing.T) {
-	const fn = "../files/StateTests/stRefundTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateBlockHash(t *testing.T) {
-	const fn = "../files/StateTests/stBlockHashTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateInitCode(t *testing.T) {
-	const fn = "../files/StateTests/stInitCodeTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateLog(t *testing.T) {
-	const fn = "../files/StateTests/stLogTests.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateTransaction(t *testing.T) {
-	const fn = "../files/StateTests/stTransactionTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestCallCreateCallCode(t *testing.T) {
-	const fn = "../files/StateTests/stCallCreateCallCodeTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestMemory(t *testing.T) {
-	const fn = "../files/StateTests/stMemoryTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestMemoryStress(t *testing.T) {
-	if os.Getenv("TEST_VM_COMPLEX") == "" {
-		t.Skip()
-	}
-	const fn = "../files/StateTests/stMemoryStressTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestQuadraticComplexity(t *testing.T) {
-	if os.Getenv("TEST_VM_COMPLEX") == "" {
-		t.Skip()
-	}
-	const fn = "../files/StateTests/stQuadraticComplexityTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestSolidity(t *testing.T) {
-	const fn = "../files/StateTests/stSolidityTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestWallet(t *testing.T) {
-	const fn = "../files/StateTests/stWalletTest.json"
-	RunVmTest(fn, t)
-}
-
-func TestStateTestsRandom(t *testing.T) {
-	fns, _ := filepath.Glob("../files/StateTests/RandomTests/*")
-	for _, fn := range fns {
-		RunVmTest(fn, t)
-	}
-}
-
-func TestVMRandom(t *testing.T) {
-	t.Skip() // fucked as of 2015-06-09. unskip once unfucked /Gustav
-	fns, _ := filepath.Glob("../files/VMTests/RandomTests/*")
-	for _, fn := range fns {
-		RunVmTest(fn, t)
-	}
 }
